:new: Add support for options.local during spawn
