+++ src/helpers.js
@@ -3,8 +3,12 @@
 /* @flow */
 
 import invariant from 'assert'
+import { async as getEnv } from 'consistent-env'
+import getNpmPath from 'sb-npm-path'
 import type { Exec$Options } from './types'
 
+const PATH_REGEX = /PATH/i
+const PATH_SEPARATOR = process.platform === 'win32' ? ';' : ':'
 export const assign = Object.assign || function(target, source) {
   for (const key in source) {
     if (source.hasOwnProperty(key)) {
@@ -39,4 +43,30 @@ export function validate(filePath: string, parameters: Array<string>, options: E
     invariant(typeof options.local === 'object', 'options.local must be an object')
     invariant(typeof options.local.directory === 'string', 'options.local.directory must be a string')
   }
+}
+
+export async function getSpawnOptions(options: Exec$Options): Promise {
+  const spawnOptions = assign({}, options, {
+    env: assign(await getEnv(), options.env)
+  })
+  const local = options.local
+  if (local) {
+    const npmPath = await getNpmPath.async(local.directory)
+    for (const key in spawnOptions.env) {
+      if (spawnOptions.env.hasOwnProperty(key) && PATH_REGEX.test(key)) {
+        const value = spawnOptions.env[key]
+        spawnOptions.env[key] = local.prepend ? npmPath + PATH_SEPARATOR + value : value + PATH_SEPARATOR + npmPath
+        break
+      }
+    }
+  }
+  spawnOptions.timeout = null
+  if (spawnOptions.env.OS) {
+    spawnOptions.env.OS = undefined
+  }
+  if (process.versions.electron) {
+    spawnOptions.env.ELECTRON_INTERNAL_RUN_AS_NODE = '1'
+    spawnOptions.env.ATOM_SHELL_INTERNAL_RUN_AS_NODE = '1'
+  }
+  return spawnOptions
 }
